+++ src/flows/user-program-management/helpers/questions.ts
@@ -20,3 +20,11 @@ export const addQuestions = [
     },
   },
 ]
+
+export const getProgramPointerInput = [
+  {
+    type: "input",
+    name: "programPointer",
+    message: "Enter the program pointer you wish to remove:",
+  },
+]